Implement on/off hours for Azure
