+++ tools/c7n_azure/c7n_azure/utils.py
@@ -16,6 +16,7 @@ import logging
 import six
 
 from azure.graphrbac.models import GetObjectsParameters, AADObject
+from azure.mgmt.resource.resources.models import GenericResource, ResourceGroupPatchable
 from msrestazure.azure_exceptions import CloudError
 from msrestazure.tools import parse_resource_id
 
@@ -116,3 +117,110 @@ class GraphHelper(object):
         elif graph_object.service_principal_names:
             return graph_object.service_principal_names[0]
         return graph_object.display_name or ''
+
+
+class TagHelper:
+
+    log = logging.getLogger('custodian.azure.utils.TagHelper')
+
+    @staticmethod
+    def update_resource_tags(tag_action, resource, tags):
+        client = tag_action.session.client('azure.mgmt.resource.ResourceManagementClient')
+
+        # resource group type
+        if tag_action.manager.type == 'resourcegroup':
+            params_patch = ResourceGroupPatchable(
+                tags=tags
+            )
+            client.resource_groups.update(
+                resource['name'],
+                params_patch,
+            )
+        # other Azure resources
+        else:
+            # generic armresource tagging isn't supported yet Github issue #2637
+            if tag_action.manager.type == 'armresource':
+                raise NotImplementedError('Cannot tag generic ARM resources.')
+
+            api_version = tag_action.session.resource_api_version(resource['id'])
+
+            # deserialize the original object
+            az_resource = GenericResource.deserialize(resource)
+
+            # create a GenericResource object with the required parameters
+            generic_resource = GenericResource(location=az_resource.location,
+                                               tags=tags,
+                                               properties=az_resource.properties,
+                                               kind=az_resource.kind,
+                                               managed_by=az_resource.managed_by,
+                                               identity=az_resource.identity)
+
+            try:
+                client.resources.update_by_id(resource['id'], api_version, generic_resource)
+            except Exception as e:
+                TagHelper.log.error("Failed to update tags for the resource.\n"
+                                    "Type: {0}.\n"
+                                    "Name: {1}.\n"
+                                    "Error: {2}".format(resource['type'], resource['name'], e))
+
+    @staticmethod
+    def remove_tags(tag_action, resource, tags_to_delete):
+        # get existing tags
+        tags = resource.get('tags', {})
+
+        # only determine if any tags_to_delete exist on the resource
+        tags_exist = False
+        for tag in tags_to_delete:
+            if tag in tags:
+                tags_exist = True
+                break
+
+        # only call the resource update if there are tags to delete tags
+        if tags_exist:
+            resource_tags = {key: tags[key] for key in tags if key not in tags_to_delete}
+            TagHelper.update_resource_tags(tag_action, resource, resource_tags)
+
+    @staticmethod
+    def add_tags(tag_action, resource, tags_to_add):
+        new_or_updated_tags = False
+
+        # get existing tags
+        tags = resource.get('tags', {})
+
+        # add or update tags
+        for key in tags_to_add:
+
+            # nothing to do if the tag and value already exists on the resource
+            if key in tags:
+                if tags[key] != tags_to_add[key]:
+                    new_or_updated_tags = True
+            else:
+                # the tag doesn't exist or the value was updated
+                new_or_updated_tags = True
+
+            tags[key] = tags_to_add[key]
+
+        # call the arm resource update method if there are new or updated tags
+        if new_or_updated_tags:
+            TagHelper.update_resource_tags(tag_action, resource, tags)
+
+    @staticmethod
+    def get_tag_value(resource, tag, enforce_utf_8=False, to_lower=False, strip_chars=[]):
+        """Get the resource's tag value."""
+
+        found = False
+        tags = resource.get('tags', {})
+
+        for key in tags.keys():
+            if key.lower() == tag.lower():
+                found = tags[key]
+                break
+
+        if found is not False:
+            if enforce_utf_8:
+                found = found.encode('utf8').decode('utf8')
+            if to_lower:
+                found = found.lower()
+            for c in strip_chars:
+                found = found.strip(c)
+        return found
